ТЗ 12 рефакторинг. 2 пулл реквест
