+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,6 +1,7 @@
 package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.controller.exceptions.IncorrectIdException;
 import ru.yandex.practicum.filmorate.controller.exceptions.ValidationException;
@@ -27,7 +28,8 @@ public class UserService {
 
     public void removeFromFriends(Integer id, Integer friendId) {
         try {
-            Optional<User> userToUpdate = userStorage.getById(id);;
+            Optional<User> userToUpdate = userStorage.getById(id);
+            ;
             friendStorage.remove(getById(id), getById(friendId));
         } catch (NullPointerException e) {
             throw new IncorrectIdException("Film для удаления не найден");
