+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/reviews/ReviewsStorage.java
@@ -13,11 +13,5 @@ public interface ReviewsStorage {
 
     void deleteReviewById(int id);
 
-    void addLikeReview(int reviewId, int userId);
-
-    void addDisLikeReview(int reviewId, int userId);
-
-    void deleteLikeReview(int reviewId, int userId);
-
     List<Review> getReviews(int filmId, int quantity);
 }
