+++ src/main/resources/schema.sql
@@ -63,3 +63,22 @@ CREATE TABLE IF NOT EXISTS PUBLIC.film_likes (
 	CONSTRAINT LIKES_FK_F FOREIGN KEY (FILM_ID) REFERENCES PUBLIC.films(FILM_ID) ON DELETE CASCADE,
 	CONSTRAINT LIKES_FK_U FOREIGN KEY (USER_ID) REFERENCES PUBLIC.users(USER_ID) ON DELETE CASCADE
 );
+DROP TABLE IF EXISTS PUBLIC.REVIEWS CASCADE;
+CREATE TABLE IF NOT EXISTS PUBLIC.REVIEWS (
+  REVIEW_ID INTEGER PRIMARY KEY AUTO_INCREMENT,
+  CONTENT VARCHAR NOT NULL,
+  IS_POSITIVE BOOLEAN NOT NULL,
+  USER_ID INTEGER NOT NULL,
+  FILM_ID INTEGER NOT NULL,
+   FOREIGN KEY (film_id) REFERENCES films (film_id),
+    FOREIGN KEY (user_id) REFERENCES users (user_id)
+);
+DROP TABLE IF EXISTS PUBLIC.LIKEREVIEW CASCADE;
+CREATE TABLE IF NOT EXISTS PUBLIC.LIKEREVIEW (
+USER_ID INTEGER NOT NULL,
+  REVIEW_ID INTEGER NOT NULL,
+  IS_LIKE BOOLEAN NOT NULL,
+  FOREIGN KEY (REVIEW_ID) REFERENCES REVIEWS (REVIEW_ID) ON DELETE CASCADE,
+    FOREIGN KEY (user_id) REFERENCES users (user_id) ON DELETE CASCADE,
+    UNIQUE (USER_ID, REVIEW_ID)
+);
